Fixed a problem with sharpen preprocessor
