+++ dataset/build.gradle
@@ -1,6 +1,5 @@
 dependencies {
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8"
-    api project(":api")
 }
