+++ examples/src/main/kotlin/examples/transferlearning/modelzoo/mobilenet/Example_2_MobileNet_additional_training.kt
@@ -41,7 +41,7 @@ private const val TRAIN_TEST_SPLIT_RATIO = 0.7
  * - All layers, excluding the last [Dense], are added to the new Neural Network, its weights are frozen.
  * - New Dense layers are added and initialized via defined initializers.
  * - Model is re-trained on [dogsCatsSmallDatasetPath] dataset.
- * - Special preprocessing (used in MobileNet during training on ImageNet dataset) is applied to images before prediction via [Sharpen] stage.
+ * - Special preprocessing (used in MobileNet during training on ImageNet dataset) is applied to images before prediction via [ImageNetize] stage.
  *
  * We use the [Preprocessing] DSL to describe the dataset generation pipeline.
  * We demonstrate the workflow on the subset of Kaggle Cats vs Dogs binary classification dataset.
@@ -69,7 +69,7 @@ fun mobilenetWithAdditionalTraining() {
         }
         transformTensor {
             sharpen {
-                modelType = ModelType.MobileNet
+                customPreprocessor = ImageNetize(modelType = ModelType.MobileNet)
             }
         }
     }
