+++ examples/src/main/kotlin/examples/transferlearning/modelzoo/resnet/resnet50/Example_2_ResNet50_prediction_additional_training.kt
@@ -68,7 +68,7 @@ fun resnet50additionalTraining() {
         }
         transformTensor {
             sharpen {
-                modelType = ModelType.ResNet_50
+                customPreprocessor = ImageNetize(ModelType.ResNet_50)
             }
         }
     }
