+++ examples/src/main/kotlin/examples/transferlearning/modelzoo/vgg19/Example_3_VGG19_additional_training.kt
@@ -72,7 +72,7 @@ fun vgg19additionalTraining() {
         }
         transformTensor {
             sharpen {
-                modelType = ModelType.VGG_19
+                customPreprocessor = ImageNetize(ModelType.VGG_19)
             }
         }
     }
